Beginning Chord Implementation
As part of my design aims, all nodes have a public-private key pair which
is to be used by higher level applications should they need to make use of
this functionality. I may add a flag to disable this mode if needed, since
key generation takes a good few seconds.