Significant changes to channel impls.
Bootstrap channels should function in a reasonble fashion now, i.e. the
server should actually map connections to ids and cleanup when needed.
This takes Chord's approach, and not my own. This may be heavier on the
server.

I also have designed it so that the bootstrap client and server send one
another their Pem keys. This is very deliberate - I intend to encrypt SDP
information soon.