Check predecessor invariant on finger table
As a node may feasibly have itself in its finger table, it follows that we
need to keep track of this invariant. Any setPredecessor calls will now
ensure that that predecessor is the largest possible entry in the finger
table - this should stop catastrophic routing failures.

Did some more thinking, and realised that cPF probably IS the best option
for message routing. I addedd an extra check to handle message when the
successor information is likely inaccurate...