Changing ChordSignalChannel to new message system.
All proxying functionality removed from the signal channel itself - this
is handled by the lower levels of the chord message core. It should now
store the last message recieved from each source, which is in turn used to
reply if it exists.

Additionally, major cleanup of the signal channel internally - plenty of
unused functions returned. I want to return to this at some point soon...