File System Authentication
Huge update. All items stored on CFS now have a concept of access rights -
a public sequence number and prior hash value combined with a private
randomised aes key. To update a file, a message is sent containing the key
and new value, along with encrypted(hash, sequence). While this can be
used to allow multiple clients to share write access to an item, this is
not recommended as they will constantly fail to have their updates
recognised due to the sequence number and hash - these are necessary steps
to mitigate the liklihood of replay attacks.

Additionally, this implements message lifetimes and keepalive messages, so
that dead nodes' data will be lost (so long as no other node on the
network is actively keeping it alive...). By design, these lifetimes are
short, and so require constant pinging.